requirements.txt
annotated-types==0.7.0
anyio==4.11.0
aiosqlite==0.19.0
asyncpg==0.29.0
black==25.9.0
certifi==2025.10.5
cfgv==3.4.0
cryptography==43.0.1
click==8.3.0
distlib==0.4.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.111.1
fastapi-cli==0.0.13
filelock==3.20.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.15
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mypy==1.18.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pre_commit==4.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
pydantic==2.12.0
pydantic-settings==2.2.1
pydantic_core==2.41.1
Pygments==2.19.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytokens==0.1.10
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
ruff==0.14.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.37.2
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.21.0
virtualenv==20.34.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
alembic==1.16.5
